requirements.txt
pandas
numpy
yfinance
requests
scikit-learn
xgboost
transformers
torch
streamlit
python-dateutil
